seperate binary for istgt integration tests
