+++ src/.gitignore
@@ -4,5 +4,6 @@
 /istgtcontrol
 /replication_test
 /mempool_test
+/istgt_integration
 *.full
 *.debug
