+++ src/Makefile.in
@@ -40,7 +40,8 @@ source   = istgt.c istgt_iscsi.c istgt_iscsi_param.c \
 	istgt_lu_ctl.c \
 	istgt_log.c istgt_conf.c istgt_sock.c istgt_misc.c \
 	istgt_queue.c istgt_crc32c.c istgt_md5.c replication.c replication_misc.c \
-	ring_mempool.c rte_ring.c data_conn.c istgt_integration_test.c mock_client.c
+	ring_mempool.c rte_ring.c data_conn.c
+
 header   = istgt_ver.h istgt.h istgt_iscsi.h istgt_iscsi_xcopy.h istgt_iscsi_param.h \
 	istgt_scsi.h istgt_proto.h istgt_lu.h \
 	istgt_log.h istgt_conf.h istgt_sock.h \
@@ -55,6 +56,9 @@ ctl_source = istgtcontrol.c istgt_conf.c istgt_log.c istgt_sock.c istgt_misc.c \
 ctl_header = istgt_ver.h istgt_conf.h istgt_log.h istgt_sock.h istgt_misc.h \
 	istgt_md5.h
 
+istgt_integration_source  = istgt_integration_test.c mock_client.c replication.c replication_misc.c rte_ring.c \
+	ring_mempool.c data_conn.c istgt_misc.c
+
 replication_test_source   = replication_test.c replication_misc.c
 replication_test_header   = replication.h istgt_integration.h
 
@@ -63,6 +67,7 @@ mempool_test_source = rte_ring.c mempool_test.c ring_mempool.c
 ISTGT    = $(source:.c=.o)
 ISTGTCONTROL = $(ctl_source:.c=.o)
 REPLICATION_TEST = $(replication_test_source:.c=.o)
+ISTGT_INTEGRATION = $(istgt_integration_source:.c=.o)
 MEMPOOL_TEST = $(mempool_test_source:.c=.o)
 
 PACKAGE_NAME = @PACKAGE_NAME@
@@ -108,6 +113,11 @@ replication_test: $(REPLICATION_TEST)
 	objcopy --only-keep-debug ${@}.full ${@}.debug
 	objcopy --strip-debug --add-gnu-debuglink=${@}.debug ${@}.full ${@}
 
+istgt_integration: $(ISTGT_INTEGRATION)
+	$(CC) $(LDFLAGS) -o ${@}.full $(ISTGT_INTEGRATION) $(LIBS)
+	objcopy --only-keep-debug ${@}.full ${@}.debug
+	objcopy --strip-debug --add-gnu-debuglink=${@}.debug ${@}.full ${@}
+
 mempool_test: $(MEMPOOL_TEST)
 	$(CC) $(LDFLAGS) -o ${@}.full $(MEMPOOL_TEST) $(LIBS)
 	objcopy --only-keep-debug ${@}.full ${@}.debug
@@ -116,7 +126,7 @@ mempool_test: $(MEMPOOL_TEST)
 build_image:
 	sh ./package.sh
 
-all: stamp-depend config.h $(SUBPROGRAM) istgtcontrol istgt replication_test mempool_test build_image
+all: stamp-depend config.h $(SUBPROGRAM) istgtcontrol istgt istgt_integration replication_test mempool_test build_image
 dev: stamp-depend config.h $(SUBPROGRAM) istgtcontrol istgt replication_test mempool_test
 
 install: install-dirs
