+++ src/mock_client.c
@@ -17,6 +17,9 @@
 typedef struct cargs_s {
 	spec_t *spec;
 	int workerid;
+	pthread_mutex_t *mtx;
+	pthread_cond_t *cv;
+	int *count;
 } cargs_t;
 
 
@@ -24,10 +27,13 @@ void create_mock_client(spec_t *);
 void *reader(void *args);
 void *writer(void *args);
 void check_settings(spec_t *spec);
-void build_cmd(cargs_t *cargs, ISTGT_LU_CMD_Ptr lu_cmd, int opcode,
+static void build_cmd(cargs_t *cargs, ISTGT_LU_CMD_Ptr lu_cmd, int opcode,
     int len);
 
-void
+cargs_t *all_cargs;
+pthread_t *all_cthreads;
+
+static void
 build_cmd(cargs_t *cargs, ISTGT_LU_CMD_Ptr cmd, int opcode,
     int len)
 {
@@ -63,21 +69,27 @@ writer(void *args)
 	cargs_t *cargs = (cargs_t *)args;
 	spec_t *spec = (spec_t *)cargs->spec;
 	int blkcnt = spec->blockcnt;
-	int blklen = spec->lu->blocklen;
+	int blklen = spec->blocklen;
 	int num_blocks = (blkcnt - 16);
 	ISTGT_LU_CMD_Ptr lu_cmd;
 	int rc, count = 0;
-	char tname[50];
 	uint64_t blk_offset, offset;
 	int len_in_blocks, len;
 	struct timespec now, start, prev;
+	pthread_mutex_t *mtx = cargs->mtx;
+	pthread_cond_t *cv = cargs->cv;
+	int *cnt = cargs->count;
 
-	snprintf(tname, 50, "mcwrite%d", cargs->workerid);
-	prctl(PR_SET_NAME, tname, 0, 0, 0);
+	clock_gettime(CLOCK_MONOTONIC, &now);
+	srandom(now.tv_sec);
+
+	snprintf(tinfo, 50, "mcwrite%d", cargs->workerid);
+	prctl(PR_SET_NAME, tinfo, 0, 0, 0);
+
 	clock_gettime(CLOCK_MONOTONIC, &start);
 	clock_gettime(CLOCK_MONOTONIC, &prev);
 
-	lu_cmd  = malloc(sizeof (ISTGT_LU_CMD));
+	lu_cmd  = (ISTGT_LU_CMD_Ptr)malloc(sizeof (ISTGT_LU_CMD));
 	memset(lu_cmd, 0, sizeof (ISTGT_LU_CMD));
 
 	while (1) {
@@ -85,7 +97,7 @@ writer(void *args)
 
 		blk_offset = random() % num_blocks;
 		offset = blk_offset * blklen;
-		len_in_blocks = random() & 15;
+		len_in_blocks = random() % 15;
 		len = len_in_blocks * blklen;
 
 		build_cmd(cargs, lu_cmd, 0, len);
@@ -99,11 +111,17 @@ writer(void *args)
 			break;
 		if (now.tv_sec - prev.tv_sec > 1) {
 			prev = now;
-			printf("wrote %d from %s\n", count, tname);
+			REPLICA_ERRLOG("wrote %d from %s\n", count, tinfo);
 		}
 	}
 end:
-	printf("wrote %d from %s\n", count, tname);
+	REPLICA_ERRLOG("exiting wrote %d from %s\n", count, tinfo);
+
+	MTX_LOCK(mtx);
+	*cnt = *cnt + 1;
+	pthread_cond_signal(cv);
+	MTX_UNLOCK(mtx);
+
 	return NULL;
 }
 
@@ -113,17 +131,22 @@ reader(void *args)
 	cargs_t *cargs = (cargs_t *)args;
 	spec_t *spec = (spec_t *)cargs->spec;
 	int blkcnt = spec->blockcnt;
-	int blklen = spec->lu->blocklen;
+	int blklen = spec->blocklen;
 	int num_blocks = (blkcnt - 16);
 	ISTGT_LU_CMD_Ptr lu_cmd;
 	int rc, count = 0;
-	char tname[50];
 	uint64_t blk_offset, offset;
 	int len_in_blocks, len;
 	struct timespec now, start, prev;
+	pthread_mutex_t *mtx = cargs->mtx;
+	pthread_cond_t *cv = cargs->cv;
+	int *cnt = cargs->count;
 
-	snprintf(tname, 50, "mcread%d", cargs->workerid);
-	prctl(PR_SET_NAME, tname, 0, 0, 0);
+	clock_gettime(CLOCK_MONOTONIC, &now);
+	srandom(now.tv_sec);
+
+	snprintf(tinfo, 50, "mcread%d", cargs->workerid);
+	prctl(PR_SET_NAME, tinfo, 0, 0, 0);
 
 	clock_gettime(CLOCK_MONOTONIC, &start);
 	clock_gettime(CLOCK_MONOTONIC, &prev);
@@ -154,16 +177,19 @@ reader(void *args)
 			break;
 		if (now.tv_sec - prev.tv_sec > 1) {
 			prev = now;
-			printf("read %d from %s\n", count, tname);
+			REPLICA_ERRLOG("read %d from %s\n", count, tinfo);
 		}
 	}
 end:
-	printf("read %d from %s\n", count, tname);
+	REPLICA_ERRLOG("exiting read %d from %s\n", count, tinfo);
+
+	MTX_LOCK(mtx);
+	*cnt = *cnt + 1;
+	pthread_cond_signal(cv);
+	MTX_UNLOCK(mtx);
+
 	return NULL;
 }
-
-cargs_t *all_cargs;
-pthread_t *all_cthreads;
 
 void
 create_mock_client(spec_t *spec)
@@ -172,8 +198,14 @@ create_mock_client(spec_t *spec)
 	int i;
 	cargs_t *cargs;
 	struct timespec now;
+	pthread_mutex_t mtx;
+	pthread_cond_t cv;
+	int count;
 
-	prctl(PR_SET_NAME, "mockclient", 0, 0, 0);
+	pthread_mutex_init(&mtx, NULL);
+	pthread_cond_init(&cv, NULL);
+
+	count = 0;
 
 	clock_gettime(CLOCK_MONOTONIC, &now);
 	srandom(now.tv_sec);
@@ -185,13 +217,20 @@ create_mock_client(spec_t *spec)
 		cargs = &(all_cargs[i]);
 		cargs->workerid = i;
 		cargs->spec = spec;
+		cargs->mtx = &mtx;
+		cargs->cv = &cv;
+		cargs->count = &count;
 		if (i < num_threads / 2)
 			pthread_create(&all_cthreads[i], NULL, &writer, cargs);
 		else
 			pthread_create(&all_cthreads[i], NULL, &reader, cargs);
 	}
-//	while (1) {
-//		sleep(1);
-//	}
+
+	MTX_LOCK(&mtx);
+	while (count != num_threads)
+		pthread_cond_wait(&cv, &mtx);
+	MTX_UNLOCK(&mtx);
+
+	return;
 }
