+++ src/replication.c
@@ -20,6 +20,7 @@
 #include "istgt_crc32c.h"
 #include "istgt_misc.h"
 #include "ring_mempool.h"
+#include "istgt_scsi.h"
 
 cstor_conn_ops_t cstor_ops = {
 	.conn_listen = replication_listen,
@@ -39,7 +40,60 @@ static void respond_with_error_for_all_outstanding_mgmt_ios(replica_t *r);
 static void inform_data_conn(replica_t *r);
 static void free_replica(replica_t *r);
 static int handle_mgmt_event_fd(replica_t *replica);
-int initialize_volume(spec_t *spec);
+int initialize_volume(spec_t *spec, int, int);
+
+#define	RCOMMON_CMD_MEMPOOL_ENTRIES	100000
+rte_smempool_t rcommon_cmd_mempool;
+size_t rcommon_cmd_mempool_count = RCOMMON_CMD_MEMPOOL_ENTRIES;
+
+#define build_rcomm_cmd 						\
+	do {								\
+		rcomm_cmd = get_from_mempool(&rcommon_cmd_mempool);	\
+		rcomm_cmd->total = 0;					\
+		rcomm_cmd->copies_sent = 0;				\
+		rcomm_cmd->total_len = 0;				\
+		rcomm_cmd->offset = offset;				\
+		rcomm_cmd->data_len = nbytes;				\
+		rcomm_cmd->state = CMD_CREATED;				\
+		rcomm_cmd->luworker_id = cmd->luworkerindx;		\
+		rcomm_cmd->mutex = 					\
+		    &spec->luworker_rmutex[cmd->luworkerindx];		\
+		rcomm_cmd->cond_var = 					\
+		    &spec->luworker_rcond[cmd->luworkerindx];		\
+		rcomm_cmd->healthy_count = spec->healthy_rcount;	\
+		rcomm_cmd->status = 0;					\
+		rcomm_cmd->completed = false;				\
+		rcomm_cmd->io_seq = ++spec->io_seq;			\
+		rcomm_cmd->state = CMD_ENQUEUED_TO_WAITQ;		\
+		switch (cmd->cdb0) {					\
+			case SBC_WRITE_6:				\
+			case SBC_WRITE_10:				\
+			case SBC_WRITE_12:				\
+			case SBC_WRITE_16:				\
+				cmd_write = true;			\
+				break;					\
+			default:					\
+				break;					\
+		}							\
+		if (cmd_write) {						\
+			rcomm_cmd->opcode = ZVOL_OPCODE_WRITE;		\
+			rcomm_cmd->iovcnt = cmd->iobufindx+1;		\
+		} else {						\
+			rcomm_cmd->opcode = ZVOL_OPCODE_READ;		\
+			rcomm_cmd->iovcnt = 0;				\
+		}							\
+		if (cmd_write) {						\
+			for (i=1; i < iovcnt + 1; i++) {		\
+				rcomm_cmd->iov[i].iov_base =		\
+				    cmd->iobuf[i-1].iov_base;		\
+				rcomm_cmd->iov[i].iov_len =		\
+				    cmd->iobuf[i-1].iov_len;		\
+			}						\
+			rcomm_cmd->total_len += cmd->iobufsize;		\
+		}							\
+	} while (0)
+
+
 
 #define BUILD_REPLICA_MGMT_HDR(_mgmtio_hdr, _mgmt_opcode, _data_len)	\
 	do {								\
@@ -330,8 +384,8 @@ update_replica_entry(spec_t *spec, replica_t *replica, int iofd)
 	rio_payload = (zvol_op_open_data_t *) malloc(
 	    sizeof (zvol_op_open_data_t));
 	rio_payload->timeout = (10 *60);
-	rio_payload->tgt_block_size = spec->lu->blocklen;
-	strncpy(rio_payload->volname, spec->lu->volname,
+	rio_payload->tgt_block_size = spec->blocklen;
+	strncpy(rio_payload->volname, spec->volname,
 	    sizeof (rio_payload->volname));
 
 	if (write(replica->iofd, rio_hdr, sizeof (*rio_hdr)) !=
@@ -421,12 +475,12 @@ send_replica_handshake_query(replica_t *replica, spec_t *spec)
 
 	mgmt_cmd = malloc(sizeof(mgmt_cmd_t));
 
-	data_len = strlen(spec->lu->volname) + 1;
+	data_len = strlen(spec->volname) + 1;
 
 	BUILD_REPLICA_MGMT_HDR(rmgmtio, mgmt_opcode, data_len);
 
 	data = (uint8_t *)malloc(data_len);
-	snprintf((char *)data, data_len, "%s", spec->lu->volname);
+	snprintf((char *)data, data_len, "%s", spec->volname);
 
 	mgmt_cmd->io_hdr = rmgmtio;
 	mgmt_cmd->io_bytes = 0;
@@ -478,12 +532,12 @@ send_replica_snapshot(spec_t *spec, replica_t *replica, char *snapname, zvol_op_
 
 	mgmt_cmd = malloc(sizeof(mgmt_cmd_t));
 	mgmt_cmd->rcomm_mgmt = rcomm_mgmt;
-	data_len = strlen(spec->lu->volname) + strlen(snapname) + 2;
+	data_len = strlen(spec->volname) + strlen(snapname) + 2;
 
 	BUILD_REPLICA_MGMT_HDR(rmgmtio, mgmt_opcode, data_len);
 
 	data = (char *)malloc(data_len);
-	snprintf(data, data_len, "%s@%s", spec->lu->volname, snapname);
+	snprintf(data, data_len, "%s@%s", spec->volname, snapname);
 
 	mgmt_cmd->io_hdr = rmgmtio;
 	mgmt_cmd->io_bytes = 0;
@@ -552,7 +606,7 @@ pause_and_timed_wait_for_ongoing_ios(spec_t *spec, int sec)
 	bool write_io_found = false;
 	replica_t *replica;
 
-	spec->lu->quiesce = 1;
+	spec->quiesce = 1;
 
 	clock_gettime(CLOCK_MONOTONIC, &last);
 	timesdiff(last, now, diff);
@@ -582,7 +636,7 @@ pause_and_timed_wait_for_ongoing_ios(spec_t *spec, int sec)
 	}
 
 	if (ret == false)
-		spec->lu->quiesce = 0;
+		spec->quiesce = 0;
 
 	return ret;
 }
@@ -669,7 +723,7 @@ int istgt_lu_create_snapshot(spec_t *spec, char *snapname, int io_wait_time, int
 	}
 	MTX_UNLOCK(&rcomm_mgmt->mtx);
 done:
-	spec->lu->quiesce = 0;
+	spec->quiesce = 0;
 	if (r == false)
 		TAILQ_FOREACH(replica, &spec->rq, r_next)
 			send_replica_snapshot(spec, replica, snapname, ZVOL_OPCODE_SNAP_DESTROY, NULL);
@@ -689,11 +743,11 @@ send_replica_status_query(replica_t *replica, spec_t *spec)
 	mgmt_cmd_t *mgmt_cmd;
 
 	mgmt_cmd = malloc(sizeof(mgmt_cmd_t));
-	data_len = strlen(spec->lu->volname) + 1;
+	data_len = strlen(spec->volname) + 1;
 	BUILD_REPLICA_MGMT_HDR(rmgmtio, mgmt_opcode, data_len);
 
 	data = (uint8_t *)malloc(data_len);
-	snprintf((char *)data, data_len, "%s", spec->lu->volname);
+	snprintf((char *)data, data_len, "%s", spec->volname);
 
 	mgmt_cmd->io_hdr = rmgmtio;
 	mgmt_cmd->io_bytes = 0;
@@ -781,9 +835,9 @@ zvol_handshake(spec_t *spec, replica_t *replica)
 		return -1;
 	}
 
-	if(strcmp(ack_data->volname, spec->lu->volname) != 0) {
+	if(strcmp(ack_data->volname, spec->volname) != 0) {
 		REPLICA_ERRLOG("volname %s not matching with spec %s volname\n",
-		    ack_data->volname, spec->lu->volname);
+		    ack_data->volname, spec->volname);
 		return -1;
 	}
 
@@ -958,7 +1012,6 @@ write_io_data(replica_t *replica, io_event_t *wevent)
 			*state = READ_IO_RESP_HDR;
 			break;
 	}
-	REPLICA_ERRLOG("got doneocount : %d : r:%p\n", donecount, replica);
 	return donecount;
 }
 
@@ -1156,6 +1209,215 @@ respond_with_error_for_all_outstanding_mgmt_ios(replica_t *r)
 	MTX_LOCK(&r->r_mtx);
 	empty_mgmt_q_of_replica(r);
 	MTX_UNLOCK(&r->r_mtx);
+}
+
+#define build_rcmd() 							\
+	do {								\
+		uint8_t *ldata = malloc(sizeof(zvol_io_hdr_t) + 	\
+		    sizeof(struct zvol_io_rw_hdr));			\
+		zvol_io_hdr_t *rio = (zvol_io_hdr_t *)ldata;		\
+		struct zvol_io_rw_hdr *rio_rw_hdr =			\
+		    (struct zvol_io_rw_hdr *)(ldata +			\
+		    sizeof(zvol_io_hdr_t));				\
+		rcmd = get_from_mempool(&rcmd_mempool);			\
+		memset(rcmd, 0, sizeof(*rcmd));				\
+		rcmd->opcode = rcomm_cmd->opcode;			\
+		rcmd->offset = rcomm_cmd->offset;			\
+		rcmd->data_len = rcomm_cmd->data_len;			\
+		rcmd->io_seq = rcomm_cmd->io_seq;			\
+		rcmd->idx = rcomm_cmd->copies_sent - 1;			\
+		rcmd->healthy_count = spec->healthy_rcount;		\
+		rcmd->rcommq_ptr = rcomm_cmd;				\
+		rcmd->status = 0;					\
+		rcmd->iovcnt = rcomm_cmd->iovcnt;			\
+		for (i=1; i < rcomm_cmd->iovcnt + 1; i++) {		\
+			rcmd->iov[i].iov_base = 			\
+			     rcomm_cmd->iov[i].iov_base;		\
+			rcmd->iov[i].iov_len = 				\
+			    rcomm_cmd->iov[i].iov_len;			\
+		}							\
+		rio->opcode = rcmd->opcode;				\
+		rio->version = REPLICA_VERSION;				\
+		rio->io_seq = rcmd->io_seq;				\
+		rio->offset = rcmd->offset;				\
+		if (rcmd->opcode == ZVOL_OPCODE_WRITE) {		\
+			rio->len = rcmd->data_len +			\
+			    sizeof(struct zvol_io_rw_hdr);		\
+			rio->checkpointed_io_seq = 0;			\
+		} else							\
+			rio->len = rcmd->data_len;			\
+		rcmd->iov_data = ldata;					\
+		rio_rw_hdr->io_num = rcmd->io_seq;			\
+		rio_rw_hdr->len = rcmd->data_len;			\
+		rcmd->iov[0].iov_base = rio;				\
+		if (rcomm_cmd->opcode == ZVOL_OPCODE_WRITE)		\
+			rcmd->iov[0].iov_len = sizeof(zvol_io_hdr_t) +	\
+			    sizeof(struct zvol_io_rw_hdr);		\
+		else							\
+			rcmd->iov[0].iov_len = sizeof(zvol_io_hdr_t);	\
+		rcmd->iovcnt++;						\
+	} while (0);							\
+
+void
+clear_rcomm_cmd(rcommon_cmd_t *rcomm_cmd)
+{
+	int i;
+	for (i=1; i<rcomm_cmd->iovcnt + 1; i++)
+		xfree(rcomm_cmd->iov[i].iov_base);
+	put_to_mempool(&rcommon_cmd_mempool, rcomm_cmd);
+}
+
+//extern void get_all_read_resp_data_chunk(void *data, struct io_data_chunk_list_t *io_data_chunk);
+//extern uint8_t * process_chunk_read_resp(struct io_data_chunk_list_t  *io_chunk_list, uint64_t len);
+
+static uint8_t *
+handle_read_consistency(rcommon_cmd_t *rcomm_cmd)
+{
+	int i;
+	uint8_t *dataptr = NULL;
+	struct io_data_chunk_list_t io_data_chunk_list;
+
+	TAILQ_INIT(&(io_data_chunk_list));
+
+	for (i = 0; i < rcomm_cmd->copies_sent; i++) {
+		if (rcomm_cmd->resp_list[i].status == 1) {
+			get_all_read_resp_data_chunk(&rcomm_cmd->resp_list[i], &io_data_chunk_list);
+		}
+	}
+
+	dataptr = process_chunk_read_resp(&io_data_chunk_list, rcomm_cmd->data_len);
+
+	return dataptr;
+}
+
+static int
+check_for_command_completion(spec_t *spec, rcommon_cmd_t *rcomm_cmd, ISTGT_LU_CMD_Ptr cmd)
+{
+	int i, rc = 0;
+	uint8_t *data = NULL;
+	int success = 0, failure = 0;
+
+	for (i = 0; i < rcomm_cmd->copies_sent; i++) {
+		if (rcomm_cmd->resp_list[i].status == 1) {
+			success++;
+		} else if (rcomm_cmd->resp_list[i].status == -1) {
+			failure++;
+		}
+	}
+
+	if (rcomm_cmd->opcode == ZVOL_OPCODE_READ) {
+		if ((rcomm_cmd->copies_sent != (success + failure))) {
+			rc = 0;
+		} else if (rcomm_cmd->copies_sent == failure) {
+			rc = -1;
+		} else {
+			data = handle_read_consistency(rcomm_cmd);
+			cmd->data = data;
+			cmd->data_len = rcomm_cmd->data_len;
+			rc = 1;
+		}
+	} else if (rcomm_cmd->opcode == ZVOL_OPCODE_WRITE) {
+		if (success >= spec->consistency_factor) {
+			rc = 1;
+			rcomm_cmd->status = 2;
+			cmd->data_len = rcomm_cmd->data_len;
+		} else if ((success + failure) == rcomm_cmd->copies_sent) {
+			rc = -1;
+			rcomm_cmd->status = -1;
+		}
+
+	}
+
+	return rc;
+}
+
+int64_t
+replicate(ISTGT_LU_DISK *spec, ISTGT_LU_CMD_Ptr cmd, uint64_t offset, uint64_t nbytes)
+{
+	int rc = -1, i;
+	bool cmd_write = false;
+	replica_t *replica;
+	rcommon_cmd_t *rcomm_cmd;
+	rcmd_t *rcmd = NULL;
+	int iovcnt = cmd->iobufindx + 1;
+	bool cmd_sent = false;
+	struct timespec abstime;
+	time_t now;
+
+	MTX_LOCK(&spec->rq_mtx);
+	if(spec->ready == false) {
+		REPLICA_LOG("SPEC is not ready\n");
+		MTX_UNLOCK(&spec->rq_mtx);
+		return -1;
+	}
+
+	build_rcomm_cmd;
+
+	TAILQ_FOREACH(replica, &spec->rq, r_next) {
+		if(replica == NULL) {
+			REPLICA_LOG("Replica not present");
+			MTX_UNLOCK(&spec->rq_mtx);
+			exit(EXIT_FAILURE);
+		}
+
+		/*
+		 * If there are some healthy replica then send read command
+		 * to all healthy replica else send read command to all
+		 * degraded replica.
+		 */
+		if (spec->healthy_rcount &&
+		    rcomm_cmd->opcode == ZVOL_OPCODE_READ) {
+			/*
+			 * If there are some healthy replica then don't send
+			 * a read command to degraded replica
+			 */
+			if (replica->state == ZVOL_STATUS_DEGRADED)
+				continue;
+			else
+				cmd_sent = true;
+		}
+
+		rcomm_cmd->copies_sent++;
+		build_rcmd();
+		put_to_mempool(&replica->cmdq, rcmd);
+		eventfd_write(replica->data_eventfd, 1);
+
+		if (cmd_sent)
+			break;
+	}
+
+	TAILQ_INSERT_TAIL(&spec->rcommon_waitq, rcomm_cmd, wait_cmd_next);
+
+	MTX_UNLOCK(&spec->rq_mtx);
+
+	// now wait for command to complete
+	while (1) {
+		MTX_LOCK(&spec->luworker_rmutex[cmd->luworkerindx]);
+		// check for status of rcomm_cmd
+		if (check_for_command_completion(spec, rcomm_cmd, cmd)) {
+			if (rcomm_cmd->status == -1)
+				rc = -1;
+			else
+				rc = cmd->data_len = rcomm_cmd->data_len;
+			put_to_mempool(&spec->rcommon_deadlist, rcomm_cmd);
+
+			MTX_LOCK(&spec->rq_mtx);
+			TAILQ_REMOVE(&spec->rcommon_waitq, rcomm_cmd, wait_cmd_next);
+			MTX_UNLOCK(&spec->rq_mtx);
+
+			MTX_UNLOCK(&spec->luworker_rmutex[cmd->luworkerindx]);
+			break;
+		}
+		now = time(NULL);
+		abstime.tv_sec = now + 1;
+		abstime.tv_nsec = 0;
+
+		pthread_cond_timedwait(&spec->luworker_rcond[cmd->luworkerindx], &spec->luworker_rmutex[cmd->luworkerindx],
+			&abstime); //timedwait
+		MTX_UNLOCK(&spec->luworker_rmutex[cmd->luworkerindx]);
+	}
+
+	return rc;
 }
 
 static void
@@ -1426,11 +1688,8 @@ initialize_replication()
 	return 0;
 }
 
-void create_mock_replicas(int, char *);
-void create_mock_client(spec_t *);
-
 int
-initialize_volume(spec_t *spec)
+initialize_volume(spec_t *spec, int replication_factor, int consistency_factor)
 {
 	int rc;
 	pthread_t deadlist_cleanup_thread;
@@ -1446,8 +1705,8 @@ initialize_volume(spec_t *spec)
 	}
 
 	spec->replica_count = 0;
-	spec->replication_factor = spec->lu->replication_factor;
-	spec->consistency_factor = spec->lu->consistency_factor;
+	spec->replication_factor = replication_factor;
+	spec->consistency_factor = consistency_factor;
 	spec->healthy_rcount = 0;
 	spec->degraded_rcount = 0;
 	spec->ready = false;
@@ -1474,9 +1733,6 @@ initialize_volume(spec_t *spec)
 	MTX_LOCK(&specq_mtx);
 	TAILQ_INSERT_TAIL(&spec_q, spec, spec_next);
 	MTX_UNLOCK(&specq_mtx);
-
-	create_mock_replicas(spec->replication_factor, spec->lu->volname);
-	create_mock_client(spec);
 
 	return 0;
 }
